fix: move matomo tracking component _app.tsx so it included on all pages
