+++ .env
@@ -7,8 +7,7 @@ NEXT_PUBLIC_MATOMO_URL_BASE=https://analytics.hel.ninja/
 NEXT_PUBLIC_MATOMO_SITE_ID=68
 NEXT_PUBLIC_SRC_URL=matomo.js
 NEXT_PUBLIC_TRACKER_URL=matomo.php
-NEXT_PUBLIC_MATOMO_ENABLED=false
-NEXT_PUBLIC_MATOMO_ENABLED=false
+NEXT_PUBLIC_MATOMO_ENABLED=true
 NEXT_PUBLIC_HEADLESS_CMS_ENABLED=true
 NEXT_PUBLIC_FORMIK_PERSIST=true
 NEXT_PUBLIC_NEWSLETTER_ENABLED=true
